# perf(matrices): Use DomainMatrix for Matrix.rref over ZZ/QQ