--- a/sympy/matrices/reductions.py
+++ b/sympy/matrices/reductions.py
@@ -1,5 +1,8 @@
 from types import FunctionType
 
+from sympy.polys.polyerrors import CoercionFailed
+from sympy.polys.domains import ZZ, QQ
+
 from .utilities import _get_intermediate_simp, _iszero, _dotprodsimp, _simplify
 from .determinant import _find_reasonable_pivot
 
@@ -242,6 +245,23 @@ def _rank(M, iszerofunc=_iszero, simplify=False):
     return len(pivots)
 
 
+def _rref_dm(dM):
+    """Compute the reduced row echelon form of a DomainMatrix."""
+    K = dM.domain
+
+    if K.is_ZZ:
+        dM_rref_z, den, pivots = dM.rref_den()
+        dM_rref_q = dM_rref_z.to_field() / den
+    elif K.is_QQ:
+        dM_rref_q, pivots = dM.rref()
+    else:
+        assert False  # pragma: no cover
+
+    M_rref = dM_rref_q.to_Matrix()
+
+    return M_rref, pivots
+
+
 def _rref(M, iszerofunc=_iszero, simplify=False, pivots=True,
         normalize_last=True):
     """Return reduced row-echelon form of matrix and indices of pivot vars.
@@ -314,13 +334,37 @@ def _rref(M, iszerofunc=_iszero, simplify=False, pivots=True,
     if you depend on the form row reduction algorithm leaves entries
     of the matrix, set ``noramlize_last=False``
     """
+    # We have to test for _rep here because there are tests that otherwise fail
+    # with e.g. "AttributeError: 'SubspaceOnlyMatrix' object has no attribute
+    # '_rep'." There is almost certainly no value in such tests. The
+    # presumption seems to be that someone could create a new class by
+    # inheriting some of the Matrix classes and not the full set that is used
+    # by the standard Matrix class but if anyone tried that it would fail in
+    # many ways.
 
-    simpfunc = simplify if isinstance(simplify, FunctionType) else _simplify
+    # Try to convert to a DomainMatrix over ZZ or QQ if possible.
+    dM = None
+    if hasattr(M, '_rep'):
+        for dom in ZZ, QQ:
+            try:
+                dM = M._rep.convert_to(dom)
+            except CoercionFailed:
+                pass
 
-    mat, pivot_cols, _ = _row_reduce(M, iszerofunc, simpfunc,
-            normalize_last, normalize=True, zero_above=True)
+    if dM is not None:
+        # Use DomainMatrix for ZZ or QQ
+        mat, pivot_cols = _rref_dm(dM)
+    else:
+        # Use the generic Matrix routine.
+        if isinstance(simplify, FunctionType):
+            simpfunc = simplify
+        else:
+            simpfunc = _simplify
+
+        mat, pivot_cols, _ = _row_reduce(M, iszerofunc, simpfunc,
+                normalize_last, normalize=True, zero_above=True)
 
     if pivots:
-        mat = (mat, pivot_cols)
-
-    return mat
+        return mat, pivot_cols
+    else:
+        return mat