--- a/sympy/matrices/reductions.py
+++ b/sympy/matrices/reductions.py
@@ -245,6 +245,36 @@ def _rank(M, iszerofunc=_iszero, simplify=False):
     return len(pivots)
 
 
+def _to_DM_ZZ_QQ(M):
+    # We have to test for _rep here because there are tests that otherwise fail
+    # with e.g. "AttributeError: 'SubspaceOnlyMatrix' object has no attribute
+    # '_rep'." There is almost certainly no value in such tests. The
+    # presumption seems to be that someone could create a new class by
+    # inheriting some of the Matrix classes and not the full set that is used
+    # by the standard Matrix class but if anyone tried that it would fail in
+    # many ways.
+    if not hasattr(M, '_rep'):
+        return None
+
+    rep = M._rep
+    K = rep.domain
+
+    if K.is_ZZ:
+        return rep
+    elif K.is_QQ:
+        try:
+            return rep.convert_to(ZZ)
+        except CoercionFailed:
+            return rep
+    else:
+        if not all(e.is_Rational for e in M):
+            return None
+        try:
+            return rep.convert_to(ZZ)
+        except CoercionFailed:
+            return rep.convert_to(QQ)
+
+
 def _rref_dm(dM):
     """Compute the reduced row echelon form of a DomainMatrix."""
     K = dM.domain
@@ -334,22 +364,8 @@ def _rref(M, iszerofunc=_iszero, simplify=False, pivots=True,
     if you depend on the form row reduction algorithm leaves entries
     of the matrix, set ``noramlize_last=False``
     """
-    # We have to test for _rep here because there are tests that otherwise fail
-    # with e.g. "AttributeError: 'SubspaceOnlyMatrix' object has no attribute
-    # '_rep'." There is almost certainly no value in such tests. The
-    # presumption seems to be that someone could create a new class by
-    # inheriting some of the Matrix classes and not the full set that is used
-    # by the standard Matrix class but if anyone tried that it would fail in
-    # many ways.
-
-    # Try to convert to a DomainMatrix over ZZ or QQ if possible.
-    dM = None
-    if hasattr(M, '_rep'):
-        for dom in ZZ, QQ:
-            try:
-                dM = M._rep.convert_to(dom)
-            except CoercionFailed:
-                pass
+    # Try to use DomainMatrix for ZZ or QQ
+    dM = _to_DM_ZZ_QQ(M)
 
     if dM is not None:
         # Use DomainMatrix for ZZ or QQ