… doc/src/modules/polys/domainmatrix.rst
… sympy/polys/matrices/tests/test_inverse.py