<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234" (see
https://tinyurl.com/auto-closing for more information). Also, please
write a comment on that issue linking back to this pull request once it is
open. -->

Makes use of fraction-free routines as discussed in gh-21834.

This based on the benchmark measurements in gh-25410.

This is built on top of the PR gh-25395 (which I had hoped could be merged before opening this PR). The sparse version of `rref_den` added there is always faster than the dense implementation added in gh-25346 and should be the main workhorse for most RREF operations.

Previous related PRs are gh-25346, gh-25369, gh-25367 and gh-25336.

#### Brief description of what is fixed or changed

This PR changes `Matrix.rref` so that for a matrix of only integers or rational numbers it will use `DomainMatrix.rref` or `DomainMatrix.rref_den` which should be entirely equivalent but a lot faster in all cases.

The `DomainMatrix.rref` method itself was also made faster for the ZZ and QQ domains by using empirically determined rules for deciding whch algorithm to use. For ZZ there are two extremes in which different algorithms are preferred: large dense matrices with large integers should use fraction-free Gauss-Jordan (GJFF) elimination over ZZ whereas very sparse matrices with small integers should use Gauss-Jordan with division (GJDIV) over QQ. For QQ the primary consideration is the nature of the denominators: small denominators should just be cleared to use GJFF over ZZ whereas very large denominators should use GJDIV over QQ. In between there are some cases where GJFF over QQ is best.

RREF is the main component of a number of matrix operations like `Matrix.inv`, `Matrix.solve`, `Matrix.nullspace`. In turn nullspace is used by `eigenvects` and `jordan_normal_form` but `eigenvects` was already previously made a lot faster by having it use DomainMatrix directly (it would benefit from using the new fraction-free routines but I haven't made any change to that here). Also `linsolve` uses `DomainMatrix.rref` directly and should be changed to use `rref_den` instead but I haven't changed that here. Also rref is used as part of the `risch` and `heurisch` integration routines although these already use `DomainMatrix` more directly: they benefit from the improvements to DomainMatrix.rref but not the fact that Matrix.rref now uses DomainMatrix.rref.

Here are some timings to illustrate the effects of this. With master the the time to invert a matrix of integers (values from 1 to 100) is:
```python
In [1]: for n in [1,2,5,10,15,16,17,18,19,20,30,40,50,100,200]:
   ...:     M = randMatrix(n)
   ...:     print()
   ...:     print(f'Time to invert a {n} by {n} matrix:')
   ...:     %time ok = M.inv()
   ...: 

Time to invert a 1 by 1 matrix:
CPU times: user 1.36 ms, sys: 0 ns, total: 1.36 ms
Wall time: 1.36 ms

Time to invert a 2 by 2 matrix:
CPU times: user 2.48 ms, sys: 0 ns, total: 2.48 ms
Wall time: 2.48 ms

Time to invert a 5 by 5 matrix:
CPU times: user 17.7 ms, sys: 0 ns, total: 17.7 ms
Wall time: 17 ms

Time to invert a 10 by 10 matrix:
CPU times: user 84.9 ms, sys: 3.7 ms, total: 88.6 ms
Wall time: 88 ms

Time to invert a 15 by 15 matrix:
CPU times: user 3.15 s, sys: 19.9 ms, total: 3.17 s
Wall time: 3.17 s

Time to invert a 16 by 16 matrix:
CPU times: user 7.41 s, sys: 39.9 ms, total: 7.45 s
Wall time: 7.45 s

Time to invert a 17 by 17 matrix:
CPU times: user 30.6 s, sys: 52 ms, total: 30.7 s
Wall time: 30.7 s

Time to invert a 18 by 18 matrix:
CPU times: user 1min 42s, sys: 108 ms, total: 1min 43s
Wall time: 1min 43s

Time to invert a 19 by 19 matrix:
CPU times: user 4min 32s, sys: 320 ms, total: 4min 32s
Wall time: 4min 32s

Time to invert a 20 by 20 matrix:
^C---------------------------------------------------------------------------
KeyboardInterrupt
```
We can see that as we go from 16x16 to 17x17 and so on each time `n` increases by 1 the runtime increases by a factor of about 3 meaning that the runtime scales exponentially like `3^n`. We can guess that the time to invert a 20x20 matrix would be something like 15 minutes but I haven't waited that long to test that.

With this PR the equivalent timings are (note that I have gmpy2 installed which affects this):
```python
In [1]: for n in [1,2,5,10,15,16,17,18,19,20,30,40,50,100,200]:
   ...:     M = randMatrix(n)
   ...:     print()
   ...:     print(f'Time to invert a {n} by {n} matrix:')
   ...:     %time ok = M.inv()
   ...: 

Time to invert a 1 by 1 matrix:
CPU times: user 1.28 ms, sys: 235 µs, total: 1.51 ms
Wall time: 1.53 ms

Time to invert a 2 by 2 matrix:
CPU times: user 958 µs, sys: 177 µs, total: 1.14 ms
Wall time: 1.15 ms

Time to invert a 5 by 5 matrix:
CPU times: user 2.77 ms, sys: 2.74 ms, total: 5.51 ms
Wall time: 5.45 ms

Time to invert a 10 by 10 matrix:
CPU times: user 15.8 ms, sys: 0 ns, total: 15.8 ms
Wall time: 15.7 ms

Time to invert a 15 by 15 matrix:
CPU times: user 21.9 ms, sys: 0 ns, total: 21.9 ms
Wall time: 21.6 ms

Time to invert a 16 by 16 matrix:
CPU times: user 23 ms, sys: 217 µs, total: 23.2 ms
Wall time: 22.8 ms

Time to invert a 17 by 17 matrix:
CPU times: user 24.7 ms, sys: 16 µs, total: 24.7 ms
Wall time: 24.7 ms

Time to invert a 18 by 18 matrix:
CPU times: user 27.2 ms, sys: 0 ns, total: 27.2 ms
Wall time: 27.2 ms

Time to invert a 19 by 19 matrix:
CPU times: user 30.6 ms, sys: 0 ns, total: 30.6 ms
Wall time: 30.6 ms

Time to invert a 20 by 20 matrix:
CPU times: user 24.7 ms, sys: 0 ns, total: 24.7 ms
Wall time: 24.8 ms

Time to invert a 30 by 30 matrix:
CPU times: user 57.9 ms, sys: 0 ns, total: 57.9 ms
Wall time: 57.9 ms

Time to invert a 40 by 40 matrix:
CPU times: user 122 ms, sys: 0 ns, total: 122 ms
Wall time: 122 ms

Time to invert a 50 by 50 matrix:
CPU times: user 216 ms, sys: 182 µs, total: 216 ms
Wall time: 216 ms

Time to invert a 100 by 100 matrix:
CPU times: user 1.67 s, sys: 12.8 ms, total: 1.68 s
Wall time: 1.68 s

Time to invert a 200 by 200 matrix:
CPU times: user 17.6 s, sys: 32.8 ms, total: 17.6 s
Wall time: 17.6 s
```
So this is 2x faster for a 2x2 matrix but that scales up to about 10000x faster for a 19x19 and probably a much bigger ration at 200x200 but no one would wait long enough to see timings for 200x200 using the existing `Matrix.rref` routine. These timings seem to scale somewhere in between `n^3` and `n^4`. I believe that asymptotically for extremely large matrices of extremely large integers this algorithm here should be `O(n^5)` if asymptotically optimal integer multiplication is used.

Note that CPython does not have asymptotically optimal multiplication whereas gmpy2 does. I think that the reasons that the existing `Matrix.rref` method are so slow as n scales up are that:

- It uses division-free rather than fraction-free Gauss-Jordan. This approach has exponential runtime in `n` even if asymptotically optimal multiplication is used.
- It uses `Rational` rather than `MPQ`. The pure Python version of `MPQ` has better arithmetic than `Rational` but when `gmpy2` is installed its `mpq` is much faster for smaller rationals but asymptotically much faster for large rationals.

Another approach that could be used is to change `Rational` (and `Integer`) to use `MPQ` and `MPZ` so that they can use gmpy2 when it is installed and also can take advantage of `PythonMPQ`'s faster arithmetic even when gmpy2 is not installed. This would speed up many things but RREF would still be slow because of the division-free approach that is used. Another possibility is that `Matrix.rref` could check to see if the matrix is all integers and then use fraction-free RREF but that is exactly what `DomainMatrix` does for us: identify an appropriate domain and then use the best algorithms for that domain.

#### Other comments

The approach here is to speed up `rref` for `Matrix` by checking to see if the matrix is all rational numbers or all integers. This is usually a very quick operation since the internal `_rep` attribute of a `Matrix` is already a `DomainMatrix` and is usually already over `ZZ` or `QQ` if the matrix is all rationals. Since matrices are mutable though it is possible that the internal `_rep` is over `EXRAW` even though its entries are all rational so we check all entries in that case. This should still be a small cost compared to the time taken to compute the actual `RREF` which is a heavy operation.

This approach could be extended to other simple domains like `RR`, `CC`, `ZZ_I`, `QQ_I`, `QQ<a>`. In the case of `QQ<a>` we need to be careful that
finding a primitive element can be slower than the rref itself so there should be some limits.

Other more interesting domains are polynomial rings and function fields like `QQ[x,y]` or `QQ(x,y)`. For these kinds of domains `rref` is fundamentally not right though and `rref_den` with clearing denominators should be used so the way to speed these up is to change methods like `inv` so that they do not attempt to use `rref` rather than speeding up `rref` itself.

This also sets the scene to be able to make use of `python_flint` to speed up matrix operations: once we know we are performing an operation like `rref` over `QQ` then the corresponding flint routines could be used which are a lot faster for dense matrices. Here are comparable timings with `python_flint`:
```python
In [1]: import flint

In [2]: to_FM = lambda M: flint.fmpq_mat([[int(e) for e in row] for row in M.tolist()])

In [3]: for n in [1,2,5,10,15,16,17,18,19,20,30,40,50,100,200]:
   ...:     M = randMatrix(n)
   ...:     print()
   ...:     print(f'Time to invert a {n} by {n} matrix:')
   ...:     fM = to_FM(M)
   ...:     %time ok = fM.inv()
   ...: 

Time to invert a 1 by 1 matrix:
CPU times: user 27 µs, sys: 1 µs, total: 28 µs
Wall time: 31.9 µs

Time to invert a 2 by 2 matrix:
CPU times: user 962 µs, sys: 51 µs, total: 1.01 ms
Wall time: 4.7 ms

Time to invert a 5 by 5 matrix:
CPU times: user 0 ns, sys: 779 µs, total: 779 µs
Wall time: 2.85 ms

Time to invert a 10 by 10 matrix:
CPU times: user 177 µs, sys: 9 µs, total: 186 µs
Wall time: 212 µs

Time to invert a 15 by 15 matrix:
CPU times: user 1.12 ms, sys: 59 µs, total: 1.17 ms
Wall time: 1.86 ms

Time to invert a 16 by 16 matrix:
CPU times: user 2.38 ms, sys: 127 µs, total: 2.51 ms
Wall time: 15.6 ms

Time to invert a 17 by 17 matrix:
CPU times: user 1.55 ms, sys: 52 µs, total: 1.6 ms
Wall time: 1.61 ms

Time to invert a 18 by 18 matrix:
CPU times: user 1.83 ms, sys: 0 ns, total: 1.83 ms
Wall time: 1.87 ms

Time to invert a 19 by 19 matrix:
CPU times: user 1.42 ms, sys: 0 ns, total: 1.42 ms
Wall time: 1.43 ms

Time to invert a 20 by 20 matrix:
CPU times: user 2.02 ms, sys: 0 ns, total: 2.02 ms
Wall time: 2.04 ms

Time to invert a 30 by 30 matrix:
CPU times: user 6.09 ms, sys: 3.97 ms, total: 10.1 ms
Wall time: 10.1 ms

Time to invert a 40 by 40 matrix:
CPU times: user 16.8 ms, sys: 0 ns, total: 16.8 ms
Wall time: 16.8 ms

Time to invert a 50 by 50 matrix:
CPU times: user 36.3 ms, sys: 0 ns, total: 36.3 ms
Wall time: 36.4 ms

Time to invert a 100 by 100 matrix:
CPU times: user 284 ms, sys: 3.87 ms, total: 288 ms
Wall time: 288 ms

Time to invert a 200 by 200 matrix:
CPU times: user 3.29 s, sys: 23.7 ms, total: 3.31 s
Wall time: 3.31 s
```
Broadly this is about 5x faster than DomainMatrix although I think that the difference widens as you go up to extremely large matrices because flint uses asymptotically optimal *matrix* multiplication and for large matrices delegates tasks like inverse to matrix multiplication. What these timings show though is that the pure Python implementation of `DomainMatrix` is not far from the performance of a highly optimised C implementation.

Here are comparable timings using `symengine` which has a C++ implementation of `Matrix`. A symengine Matrix is like a sympy Matrix in that it allows arbitrary expressions in the matrix and uses its equivalent of `Expr` rather than choosing particular domains like `python_flint` or `DomainMatrix`. This means that it is not as fast when there is a well defined domain as in this case but otherwise I am not sure what algorithm it uses for `inv` or `rref`:
```python
In [1]: import symengine

In [2]: to_SE = lambda M: symengine.Matrix([[int(e) for e in row] for row in M.tolist()])

In [3]: for n in [1,2,5,10,15,16,17,18,19,20,30,40,50,100,200]:
   ...:     M = randMatrix(n)
   ...:     print()
   ...:     print(f'Time to invert a {n} by {n} matrix:')
   ...:     sM = to_SE(M)
   ...:     %time ok = sM.inv()
   ...: 

Time to invert a 1 by 1 matrix:
CPU times: user 41 µs, sys: 5 µs, total: 46 µs
Wall time: 50.5 µs

Time to invert a 2 by 2 matrix:
CPU times: user 44 µs, sys: 5 µs, total: 49 µs
Wall time: 53.6 µs

Time to invert a 5 by 5 matrix:
CPU times: user 212 µs, sys: 0 ns, total: 212 µs
Wall time: 216 µs

Time to invert a 10 by 10 matrix:
CPU times: user 4.29 ms, sys: 0 ns, total: 4.29 ms
Wall time: 4.31 ms

Time to invert a 15 by 15 matrix:
CPU times: user 13.6 ms, sys: 162 µs, total: 13.8 ms
Wall time: 13.4 ms

Time to invert a 16 by 16 matrix:
CPU times: user 18.6 ms, sys: 0 ns, total: 18.6 ms
Wall time: 18.5 ms

Time to invert a 17 by 17 matrix:
CPU times: user 19.8 ms, sys: 0 ns, total: 19.8 ms
Wall time: 19.8 ms

Time to invert a 18 by 18 matrix:
CPU times: user 23.7 ms, sys: 0 ns, total: 23.7 ms
Wall time: 23.8 ms

Time to invert a 19 by 19 matrix:
CPU times: user 26.8 ms, sys: 0 ns, total: 26.8 ms
Wall time: 26.9 ms

Time to invert a 20 by 20 matrix:
CPU times: user 32.9 ms, sys: 0 ns, total: 32.9 ms
Wall time: 32.9 ms

Time to invert a 30 by 30 matrix:
CPU times: user 131 ms, sys: 0 ns, total: 131 ms
Wall time: 131 ms

Time to invert a 40 by 40 matrix:
CPU times: user 386 ms, sys: 0 ns, total: 386 ms
Wall time: 386 ms

Time to invert a 50 by 50 matrix:
CPU times: user 866 ms, sys: 2.72 ms, total: 869 ms
Wall time: 870 ms

Time to invert a 100 by 100 matrix:
CPU times: user 12.5 s, sys: 2.05 ms, total: 12.5 s
Wall time: 12.5 s

Time to invert a 200 by 200 matrix:
CPU times: user 3min 24s, sys: 95.2 ms, total: 3min 24s
Wall time: 3min 24s
```
So SymEngine is faster for very small matrices but becomes slower than DomainMatrix as we scale up to very large matrices. My guess is that these differences are explained by the fact that at small sizes pure Python overhead slows down DomainMatrix but at larger sizes the effect of an asymptotically better algorithm (fraction-free Gauss Jordan) means that DomainMatrix is faster.

#### Release Notes

<!-- Write the release notes for this release below between the BEGIN and END
statements. The basic format is a bulleted list with the name of the subpackage
and the release note for this PR. For example:

* solvers
  * Added a new solver for logarithmic equations.

* functions
  * Fixed a bug with log of integers. Formerly, `log(-x)` incorrectly gave `-log(x)`.

* physics.units
  * Corrected a semantical error in the conversion between volt and statvolt which
    reported the volt as being larger than the statvolt.

or if no release note(s) should be included use:

NO ENTRY

See https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more
information on how to write release notes. The bot will check your release
notes automatically to see if they are formatted correctly. -->

If gh-25395 is not merged before this then the release notes from there should be included here if this is merged.

<!-- BEGIN RELEASE NOTES -->
* matrices
    * The DomainMatrix.rref and rref_den methods have been optimised for the ZZ and QQ domains. The new methods for ZZ select between using division or fraction-free Gauss-Jordan elimination depending on sparseness, bit counts, and matrix dimensions. For QQ the denominators  are checked to consider whether to clear denominators and use ZZ or otherwise use division over QQ.
    * `Matrix.rref` now uses `DomainMatrix.rref` when the matrix consists only of integers or rational numbers. This makes the `rref` operation a lot faster and speeds up many other matrix operations that use `rref` such as computing a matrix inverse, nullspace or solving a matrix equation. The exact timing difference depends on a number of factors and is larger for larger matrices. As an example computing the inverse of a dense 19x19 matrix of integers values between 0 and 100 is about 10000x faster. The speed difference is bigger if gmpy2 is installed since that speeds up elementary integer and rational arithmetic and will now be used by `Matrix.rref` when possible.
<!-- END RELEASE NOTES -->