@oscarbenjamin 2023-07-29
> This is built on top of the PR [gh-25395](https://github.com/sympy/sympy/pull/25395) (which I had hoped could be merged before opening this PR). The sparse version of `rref_den` added there is always faster than the dense implementation added in [gh-25346](https://github.com/sympy/sympy/pull/25346) and should be the main workhorse for most RREF operations.

Note that the diff here would be a lot smaller if gh-25395 was merged. I wanted this to be two separate PRs where one adds the algorithm for sparse fraction-free elimination to DomainMatrix and then this PR adds the code for selecting the fastest algorithm and for making Matrix.rref use DomainMatrix.rref.